# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Âu Lạc Railway API v1
  version: "1.0.0"
  description: >
    MVP demo API — leg-splitting, gap-merging, flexible pricing.
    Source of truth per plan/00_MASTER_PLAN.md §7 và docs/API_Contract.md.
    Money = int64 VND, rounded to 1000. Timestamps = ISO-8601 UTC on the demo
    scenario clock (2026-06-15), not wall time. segment_id is 1-based (L1..L7).
servers:
  - url: /api/v1
paths:
...
  /holds:
    post:
      summary: Atomic hold on the offer's seat plan
      operationId: createHold
      tags: [booking]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HoldRequest" }
            example:
              offer_id: offer_9981
              expected_matrix_version: 5
              passenger_name: "Nguyễn Văn A"
      responses:
        "201":
          description: Hold active
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HoldResponse" }
        "422":
          description: CONSENT_REQUIRED — seat_plan có >=2 leg (ghép nhiều ghế) nhưng consent != true
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
              example:
                error_code: CONSENT_REQUIRED
                message: "Phương án ghép nhiều ghế cần khách xác nhận đồng ý đổi chỗ trước khi giữ ghế"
                details: { offer_id: offer_9981, so_lan_doi_cho: 1 }
        "409": { $ref: "#/components/responses/Conflict" }
        "410": { $ref: "#/components/responses/Gone" }
  /bookings/{hold_id}/confirm:
    post:
      summary: Confirm hold (HELD -> SOLD), never re-prices, idempotent
      operationId: confirmBooking
      tags: [booking]
      parameters:
        - name: hold_id
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfirmResponse" }
        "410": { $ref: "#/components/responses/Gone" }
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Conflict:
      description: "409 STALE_SNAPSHOT | SEAT_CONFLICT"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Gone:
      description: "410 OFFER_EXPIRED | HOLD_EXPIRED"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    HoldStatus:
      type: string
      enum: [ACTIVE, CONFIRMED, EXPIRED, CANCELLED]
    ErrorCode:
      type: string
      enum:
        - NO_SAME_SEAT_OPTION
        - SOLD_OUT_TRUE
        - ALLOCATION_REJECTED
        - STALE_SNAPSHOT
        - SEAT_CONFLICT
        - OFFER_EXPIRED
        - HOLD_EXPIRED
        - POLICY_UNAVAILABLE
        - CONSENT_REQUIRED
        - RESOURCE_NOT_FOUND
        - FORBIDDEN
        - GUARDRAIL_VIOLATION
    ErrorEnvelope:
      type: object
      required: [error_code, message]
      properties:
        error_code: { $ref: "#/components/schemas/ErrorCode" }
        message: { type: string }
        details: { type: object }
...
    HoldRequest:
      type: object
      required: [offer_id, expected_matrix_version]
      properties:
        offer_id: { type: string }
        expected_matrix_version: { type: integer }
        passenger_name: { type: string }
        consent:
          type: boolean
          default: false
          description: Bắt buộc true khi offer.requires_customer_consent — thiếu -> 422 CONSENT_REQUIRED (P5).
    HoldResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            hold_id: { type: string }
            status: { $ref: "#/components/schemas/HoldStatus" }
            expires_at: { type: string, format: date-time }
            new_matrix_version: { type: integer }
    ConfirmResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            booking_id: { type: string }
            status: { type: string }
            final_price_vnd: { type: integer }
            decision_record_id: { type: string }
...